test: TravelogueRequestBuilder와 TravelogueDayRequestBuilder 별도 클래스로 분리
