+++ backend/src/test/java/kr/touroot/travelogue/helper/TravelogueRequestBuilder.java
@@ -3,12 +3,9 @@ package kr.touroot.travelogue.helper;
 import java.util.ArrayList;
 import java.util.List;
 import kr.touroot.travelogue.dto.request.TravelogueDayRequest;
-import kr.touroot.travelogue.dto.request.TraveloguePlaceRequest;
 import kr.touroot.travelogue.dto.request.TravelogueRequest;
 import kr.touroot.travelogue.fixture.TravelogueDayFixture;
 import kr.touroot.travelogue.fixture.TravelogueFixture;
-import kr.touroot.travelogue.fixture.TraveloguePhotoFixture;
-import kr.touroot.travelogue.fixture.TraveloguePlaceFixture;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 
@@ -40,28 +37,5 @@ public class TravelogueRequestBuilder {
 
     void addDayRequest(TravelogueDayRequest dayRequest) {
         dayRequests.add(dayRequest);
-    }
-
-    @RequiredArgsConstructor
-    public static class TravelogueDayRequestBuilder {
-        private final TravelogueRequestBuilder parentBuilder;
-        private final TravelogueDayFixture travelogueDayFixture;
-        private final List<TraveloguePlaceRequest> placeRequests = new ArrayList<>();
-
-        public TravelogueDayRequestBuilder addPlaceWithPhotosIntoDay(
-                TraveloguePlaceFixture traveloguePlaceFixture,
-                List<TraveloguePhotoFixture> traveloguePhotoFixtures
-        ) {
-            placeRequests.add(traveloguePlaceFixture.getCreateRequestWith(traveloguePhotoFixtures));
-            return this;
-        }
-
-        public TravelogueRequestBuilder buildDay() {
-            TravelogueDayRequest dayRequest = travelogueDayFixture.getCreateRequestWith(
-                    placeRequests.toArray(new TraveloguePlaceRequest[0])
-            );
-            parentBuilder.addDayRequest(dayRequest);
-            return parentBuilder;
-        }
     }
 }
